feat: support custom container handler class for `@Secret`
